made search bar a separate component from HomeScreen.js
