--- SearchBar.js
@@ -1,22 +1,0 @@
-import React from 'react';
-import {View, Text, StyleSheet} from 'react-native'
-
-function SearchBar() {
-    return (
-        <View style={styles.container}>
-            <Text> To be developed </Text>
-        </View>
-
-    );
-}
-
-const styles = StyleSheet.create({
-    container: {
-        flex: 1,
-        justifyContent: 'center',
-        alignItems: 'center',
-        backgroundColor: '#FFF4BE'
-    },
-})
-
-export default SearchBar
